fix(patina): allow type-only exports in script setup
